# Optimize and reduce memory usage of XML serialization