@nielsdos 2024-05-11
> How looks like the perf improvement ? :)

```
[niels@ARCH RELx64]$ hyperfine './sapi/cli/php xml.php' './sapi/cli/php_old xml.php'
Benchmark 1: ./sapi/cli/php xml.php
  Time (mean ± σ):     258.7 ms ±   5.9 ms    [User: 250.2 ms, System: 6.6 ms]
  Range (min … max):   252.1 ms … 267.9 ms    11 runs
 
Benchmark 2: ./sapi/cli/php_old xml.php
  Time (mean ± σ):     319.7 ms ±  19.1 ms    [User: 281.9 ms, System: 36.1 ms]
  Range (min … max):   301.9 ms … 368.9 ms    10 runs
 
Summary
  ./sapi/cli/php xml.php ran
    1.24 ± 0.08 times faster than ./sapi/cli/php_old xml.php
```

As for memory usage:
```
Before: Maximum resident set size (kbytes): 25880
After: Maximum resident set size (kbytes): 24568
```